PROJECT/backend/requirements.txt
google-generativeai==0.7.2
Flask==3.0.3
google-search-results==2.4.2
Flask-Cors>=4.0.1
Werkzeug>=2.3.7  
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.43
langchain-groq==0.1.10
langchain-google-genai==1.0.8
python-dotenv==1.0.1
Flask-SocketIO>=5.4.1
spacy==3.7.1
pypdf==5.1.0
chromadb==0.5.23
faiss-cpu==1.10.0
sentence-transformers===3.4.1
jq==1.8.0
hf_xet==1.0.3
playwright==1.50.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
crawl4ai==0.4.247
cohere==5.8.1
yfinance==0.2.58
requests==2.32.3
fastapi==0.115.0
uvicorn==0.30.6
serpapi==0.1.5
